tests for GNNlib (#466)
