+++ GNNlib/Project.toml
@@ -20,8 +20,8 @@ CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 GNNlibCUDAExt = "CUDA"
 
 [compat]
-ChainRulesCore = "1.24"
 CUDA = "4, 5"
+ChainRulesCore = "1.24"
 DataStructures = "0.18"
 GNNGraphs = "1.0"
 LinearAlgebra = "1"
@@ -32,7 +32,10 @@ Statistics = "1"
 julia = "1.10"
 
 [extras]
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["Test", "ReTestItems", "Reexport", "SparseArrays"]
