+++ GNNlib/test/runtests.jl
@@ -1,0 +1,6 @@
+using GNNlib
+using Test
+using ReTestItems
+using Random, Statistics
+
+runtests(GNNlib)
