+++ GNNlib/test/shared_testsetup.jl
@@ -1,0 +1,12 @@
+@testsetup module SharedTestSetup
+
+import Reexport: @reexport
+
+@reexport using GNNlib
+@reexport using GNNGraphs
+@reexport using NNlib
+@reexport using MLUtils
+@reexport using SparseArrays
+@reexport using Test, Random, Statistics
+
+end
